WIN-99 restore Schedule lazy split integrity
